Update to pymc5, for python 3.11 compatibility (#33)
* Update to pymc5, for python 3.11 compatibility

Currently, pymc is pinned to v4. However, this in turn prevents from updating to python3.11. pymc5 requires a dropin replacement of aesara-->pytensor, as referenced [here](https://discourse.pymc.io/t/pymc-4-aesara-cannot-convert-elemwise-pow-no-inplace-0-to-a-tensor-variable/11436/2), but is otherwise entirely compatible with ringdown in its present form. Having made this change I tested with the GW150914.ipynb standard example notebook, and found everything to be working fine.

* pm.distributions.transforms.ordered --> Ordered

* multivariate_ordered

A correction to the previous commit - multivariate_ordered is the correct corresponding object to "ordered" from pymc4

* Import json, and reference json.JSONDecodeError

* Fix typo in variable name

* Add missing imports

---------

Co-authored-by: Rhiannon Udall <rudall@caltech.edu>